requirements.txt
numpy
pandas 
matplotlib
seaborn
opencv-python
tensorflow
streamlit
pillow
webcolors